Clean part 6 with working conditionals and Included test cases
Cleaned up the code a significant amount and added auxiliary functions for EVAL.
This is a fully functioning part 6 but need clarification on how read function is intended to work before implementation.
///////////
Have included thoughts bookmark which is a pseudo readmead that indicates what questions I am currently asking myself and what I need
to work on currently. This commit signifies that I will begin working on part 7 officially now.